+++ src/authentication/controller.js
@@ -1,4 +1,4 @@
-import { CustomError, asyncWrapper } from "../../utils/index.js";
+import { CustomError, asyncWrapper, emailVerification } from "../../utils/index.js";
 import Users from "../../models/users.js";
 import { StatusCodes } from "http-status-codes";
 import bcrypt from "bcrypt";
@@ -6,7 +6,7 @@ import Cart from "../../models/cart.js";
 import jwt from "jsonwebtoken";
 import "dotenv/config.js";
 import RefreshToken from "../../models/refreshToken.js";
-import nodemailer from "nodemailer";
+
 
 // generate token function
 export const GenerateTokenFunction = (userId, type) => {
@@ -22,59 +22,6 @@ export const GenerateTokenFunction = (userId, type) => {
     console.log(error.message);
   }
 };
-
-// email sender authentication
-let transporter = nodemailer.createTransport({
-  service: "gmail",
-  secure: true,
-  auth: {
-    type: "OAuth2",
-    user: process.env.EMAIL_USER,
-    pass: process.env.EMAIL_PASS,
-    clientId: process.env.CLIENT_ID,
-    clientSecret: process.env.CLIENT_SECRET,
-    accessToken: process.env.ACCESS_TOKEN,
-    refreshToken: process.env.REFRESH_TOKEN,
-  },
-});
-
-// email verification function
-const emailVerification = asyncWrapper(async (user) => {
-  // generate verification token with user_id
-  let token = jwt.sign({ userId: user.user_id }, process.env.JWT_SECRET, {
-    expiresIn: "10m",
-  });
-  let link = `http://localhost:3000/verify/${token}`;
-  let message = {
-    from: "freakydmuse@gmail.com",
-    to: user.email,
-    subject: "E-commerce App: Welcome to E-commerce App",
-    html: `<p>Welcome to E-commerce App ${user.first_name + ' ' + user.last_name}</p><br/><p>Here is your verification link, click <a href=${link} >here<a/> to verify. The link expires in 10 minutes.</p>`,
-  };
-
-  try {
-    const response = await transporter.sendMail(message);
-    
-  } catch (error) {
-    
-    throw new CustomError('Sorry an error occurred while registering, please try again')
-  }
-});
-
-// verify token from email
-export const emailTokenVerification = asyncWrapper(async (req, res) => {
-  const { token } = req.params;
-  
-  if(!token){
-    throw new CustomError('Sorry an error occurred, could not parse token');
-  }
-
-  const decoded = jwt.verify(token, process.env.JWT_SECRET);
-  let userId = decoded.userId;
-  await Users.verifyEmail(userId);
-  
-  res.send('Account verified successfully');
-});
 
 // create user
 export const signUp = asyncWrapper(async (req, res) => {
