+++ utils/index.js
@@ -2,3 +2,4 @@ export { default as globalErrorHandler } from './globalErrorHandler.js';
 export { CustomError, ValidationError } from './customError.js';
 export { default as asyncWrapper } from './aysncWrapper.js';
 export { default as redisStore } from './sessionStorage.js'
+export { default as emailVerification } from './emailVerification.js'
